recommend-server/requirements.txt
python-dotenv
Flask
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.3
pyspark
findspark
numpy
pandas
# pip install -r requirements.txt